+++ packages/terra-aggregator/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Changed
+  * Migrated docs to terra-framework-docs
+
 ## 4.70.0 - (July 13, 2021)
 
 * Changed